move admin view tests to separate file
